extracted base class methods
